Only support channel requests in their respective channels
